json_transport: support custom errors
Custom errors are provided in form of map from error name to error sample
object. If the error belongs to one of the types, it is encoded to JSON
and passed in a separate JSON field ("detail") as well as its type (in
JSON field "code"). If the client encounters the field, it tries
to parse the error into the object of this type.

If an error of a registered type is wrapped (e.g. using fmt.Errorf) then
it is unwrapped and put into "detail".

Unregistered errors are passed as before.